Don't show drop hint on same device
Moves where we prevent the default behavior to hint that dropping is not allowed on the same side of the transfer window. This doesn't actually result in any behavior change (you couldn't drop to/from the same device before) but it no longer shows the "+" icon while dragging on the same side of the screen.